--story=116941450 环境管理部分代码优化2
feat:新增节点时，写入节点svrId

+++ src/backend/ci/ext/tencent/common/common-devcloud/src/main/kotlin/com/tencent/devops/common/environment/agent/pojo/agent/RawCmdbNode.kt
@@ -35,5 +35,5 @@ data class RawCmdbNode(
     val displayIp: String,
     val osName: String,
     var agentStatus: Boolean,
-    val serverId: Int
+    val serverId: Long
 )
